requirements.txt
# Core ML and NLP dependencies
transformers==4.38.2  # For expert models and fine-tuning (google/mt5-small)
torch==2.2.1  # Backend for transformers and custom ML tasks
sentence-transformers==2.5.1  # For diversity scoring in Diversity Enhancer Engine
scikit-learn  # For clustering/classification in future engines

# Mathematical and scientific computing
sympy==1.12  # For Calculator Engine (symbolic and numerical math)
numpy==1.26.4  # For numerical computations in Code Interpreter and Graphing Engines
scipy==1.12.0  # For future simulation or optimization engines

# Visualization
matplotlib==3.8.3  # For Graphing Engine (plots, charts)
seaborn==0.13.2  # For enhanced visualizations in future plotting
plotly==5.20.0  # For interactive plots in future web UIs

# Data management
psycopg2-binary==2.9.9  # For potential PostgreSQL integration
pandas==2.2.1  # For data manipulation and analysis

# Code execution and security
restrictedpython==7.0  # For safe Python execution in Code Interpreter Engine
micropip==0.5.0  # For Pyodide package management in browser (note: Pyodide setup separate)

# Web and API interactions
requests==2.31.0  # For Web Search Engine (mock API, replace with SerpAPI)
aiohttp==3.9.3  # For asynchronous API calls in future tools

# Document processing
pylatex==1.4.2  # For programmatic LaTeX generation in LaTeX Formatting Engine
Jinja2==3.1.3  # For template rendering in future document/HTML outputs

# Logging and monitoring
loguru==0.7.2  # For enhanced logging in production

# Performance and scalability
redis==5.0.3  # For caching (e.g., database results, plot PNGs)
celery==5.3.6  # For distributed task queuing in future async tasks

# Testing and development
pytest==8.1.1  # For unit testing engines and integrations
black==24.3.0  # For code formatting
flake8==7.0.0  # For linting and code quality

# Future tools
beautifulsoup4==4.12.3  # For web scraping in future data extraction
spacy==3.7.4  # For NLP in future summarization/query expansion
lxml==5.1.0  # For MathML/XML processing in future converters
networkx==3.2.1  # For graph-based analysis in future simulation engines
tqdm==4.66.2  # For progress bars in data processing
pyyaml==6.0.1  # For configuration file parsing in future setups